requirements.txt
tabula-py==2.9.0
pandas==2.2.1
fastapi==0.110.0
mangum==0.17.0
python-multipart==0.0.9
